recipe-backend/requirements/dev.txt
-r prod.txt
flake8==3.7.9
pytest-django==3.8.0
